requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.5
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.4
psycopg2-binary==2.9.9
aiosqlite==0.19.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
jinja2==3.1.2

# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# sqlalchemy==2.0.23
# alembic==1.12.1
# pydantic[email]==2.5.0
# python-multipart==0.0.6
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# python-dotenv==1.0.0
# openai==1.3.5
# sentence-transformers==2.2.2
# scikit-learn==1.3.2      # ✅ Wheel available for Windows/macOS/Linux
# numpy==1.24.4            # ✅ Wheel available (use older version for compatibility)
# psycopg2-binary==2.9.9   # ✅ Already a wheel (binary version)
# aiosqlite==0.19.0
# httpx==0.25.2
# pytest==7.4.3
# pytest-asyncio==0.21.1
# requests==2.31.0
# jinja2==3.1.2

